+++ DesignToolbox/DesignToolbox/Pages/Components/Controls/TextArea/TextAreaConfigurationView.swift
@@ -310,7 +310,7 @@ struct TextAreaConfigurationView: View {
                                    chips: CircularProgressIndicatorConfigurationModel.Variant.chips)
 
                     if configurationModel.progressVariant == .determinate {
-                        ProgressControl(progress: $configurationModel.progressValue)
+                        DesignToolboxProgressControl(progress: $configurationModel.progressValue)
                     }
                 }